Merge r.watershed2 from Markus Metz into r.watershed. (merge from devbr6 r34645)
fast version of r.watershed:
Implemented is a new sorting algorithm for both the ram and the 
segmented version, resulting in significant speed improvements,
with identical results.

The new sorting algorithm is implemented for both the ram and the
segmented version.
The new segmented version is still much slower than the new ram 
version, but the new segmented version is already much faster than 
the original ram version. A new option is available for the segmented 
mode specifying the maximum amount of memory to be used in MB.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@34646 15284696-431f-4ddb-bdfa-cd5b030d7da7