testing: remove note about specific test server (introduced in https://trac.osgeo.org/grass/changeset/62358)
The documentation does not document specific how tests are executed at specific machine but it is a general description how things works. The file contains example how to set up a cron job using a (example) shell script which wraps several steps together. There can (and should) be (any) number of machines running the test and different other checks besides the fact that developers should run the tests on their machines before the commit for example using a script similar to one provided.

git-svn-id: https://svn.osgeo.org/grass/grass/trunk@62362 15284696-431f-4ddb-bdfa-cd5b030d7da7