wxGUI/toolboxes: make tooboxes a little bit more independent from GUI
This uses various workarounds, lazy imports and wrappers to not import anything from wxGUI (and thus wx) or to import it only if it is really needed. This avoids issues when building toolboxes during compilation and the environment is not set correctly and wx (and thus wxGUI) cannot be imported properly. This does not make GUI toolboxes mechanism independent on wxGUI (https://trac.osgeo.org/grass/ticket/2142). It just makes compilation more robust and shows which dependencies must be solved to create standalone toolboxes mechanism. See also https://trac.osgeo.org/grass/ticket/1819.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@64677 15284696-431f-4ddb-bdfa-cd5b030d7da7