docs: do not add small topics to index
Topics with less than 3 modules in them are not included in Topics index. This makes the index shorter and easier to navigate. Topics are still included in keyword index as keywords, so there is no need to list them when they are not so important (assuming importance/impact equals module count).

The topic pages are still generated, so there is no need to change anything for the modules. By clicking at the second keyword users still gets to the topic page. It is just not advertised in the Topics index.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@66040 15284696-431f-4ddb-bdfa-cd5b030d7da7