r.in.lidar: vector output with grid-based decimation [news]
Creates one point per raster cell. Point gets
XYZ which are mean of all points in the cell.
No categories are created or preserved as it
would be unclear what to take (we cannot average
retrun number).

Some improvements are needed for history and done
message code. Vector writting could be more separated.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@67210 15284696-431f-4ddb-bdfa-cd5b030d7da7