r.sim.water and r.sim.sediment: do not use lib globals outside of the lib
Using library global variables outside of a dynamically linked
library may not be portable (MS Windows). See #2966.

Use structure and a setter function to set all the library
variables. Some other calls and structures needed to be
moved around to not use the library globals from main
functions of individual modules. New header added which
is the interface for the SIMWE library.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@68320 15284696-431f-4ddb-bdfa-cd5b030d7da7