v.in.lidar, v.in.pdal: add -c to not store unique cat for each point
No default and explicit flag works better in GUI forms
which do not allow passing nothing when there is a default answer.
The command line is also better as it requires -c or id_layer
when other layer options are used.

It also checks if two layer options are set to
the same numbers and fatals if they are
(allowed in GRASS but does not make sense here).

Verbose message when unique category is stored
for each point.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@68476 15284696-431f-4ddb-bdfa-cd5b030d7da7