add arrow symbol to basic symbols
Derived from geology/half-arrow_left. It could be part of geology,
n_arrows (makes simple N arrow) or extra (there are already 3 arrows in basic).
Name could be also simple_arrow (but arrow is a simple name).
There could be more versions with different lenght or angle of the lines.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@70794 15284696-431f-4ddb-bdfa-cd5b030d7da7