init: do not show non-standard help options
We are now not showing -help (legacy) and --h (module-like) in the same way
we were not showing help (legacy from modules and subcommand-like).

They are still supported for convenience, but not advertised.

This also simplifies the help text a little bit. (Some tools
do not show --help etc. in the help text at all.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73101 15284696-431f-4ddb-bdfa-cd5b030d7da7