init: do not remove wrong mapset .tmp, rely on clean_temp (see https://trac.osgeo.org/grass/ticket/3635)
The previous behavior was deleting whole mapset temporary directory
of a mapset set at the startup. When started in mapset A, switched to B,
exited, whole .tmp of A was removed, B was cleaned by clean_temp.
Now cleaning of A is done during switching (as before) but it is not
touched again, so no running processed in a (new) session in A are
in denger of loosing temporary files. The mapset active at exit is
cleaning using clean_temp as before, but only by that.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73336 15284696-431f-4ddb-bdfa-cd5b030d7da7