v.in.ascii: include values into error messages (#104)
Not only the column, but also the value is now part of
the message about type since the value influenced the type
which makes it easier to find the issue. It reports the value for
the main columns such as x and y, and also for any mismatch
in requested and actual value types.

It stores first encountered value as sample for
all types and stores NULL for empty columns.
Type is used for (other) attribute columns type errors
while NULLs to avoid message about integers in data
given for empty columns which was confusing.

Uses single quotes around data which is according to
the Submitting rules (MessageStandardization) which require
single quotes around strings, but does not apply rules otherwise
to be consistent(ly wrong together) with most of v.in.ascii.

Also unify function parameters and line breaks.