Module level import not at top of file (E402)
Fixing by actually moving imports (hashlib), moving the non-importing code down
after the imports (matplotlib.use and Python 2/3 variables).

Several files ignored using the per-file-ignores. These may or may not be substantiated as exceptions,
so keeping them in the config file rather than in the file itself to have here a clear warning
that the check was disabled.

In general E402 should be fixable by moving code around and couple inline ignores like in the case
of matplotlib.use in gunittest/multireport.