Fix Flake8 errors in temporal modules (#573)
* Comparison to None should be 'if cond is not None:' (E711)
* Comparison to False should be 'if cond is False:' or 'if not cond:' (E712)
* Imported but unused (F401)
* Do not use bare 'except' (E722)

There is the same or similar code in several modules checking presence of ply which is now documented as unused and lines are ignored for Flake8. Catching failed test import of ply using ImportError not bare except.

In some cases, the `is` operator is used for comparison (`check is False`), while in almost all other `cases == False` is simply removed (`not check`). The goal is to make less assumptions about the values coming from the "check" function and in the check variable. When the function name and variable name are something like `is_ok()` and `ok`, then it is pretty clear the possible values are True and False. With check, as a reader of this code, it is not clear if `check == False` and `not check` are really the same, because check could be e.g. None if the check failed and the case should be handled in some other way. (Likely not the case here, but we are already using same tactics in GUI.) With `is_ok()` and `ok` it is safer to make the assumption about what the possible values are. Both `not check` and `check is False` are allowed, so using here whatever is safer and easier to review. Couple other places have `is False` where it is not clear that the expression is boolean.

The errors are now reordered: serious first, whitespace last.