Add --no-lock, --no-clean, and --clean-only
This allows to skip locking of a mapset which in combination with no cleaning
of a mapset allows for multiple jobs to be executed in one mapset using --exec.
This should be the same as running multiple jobs together right in a single session
(having the same limitations, but also not touching any general files).

To account for the use case when the mapset is not used for any interactive session,
a separate flag for cleaning is provided for convenience (same as something like
--exec g.region -p).

Similarly to --tmp-location and --tmp-mapset, usage of --no-lock and --no-clean without
--exec is not disallowed, but it is not documented either before its usefulness is
evaluated. One use case is starting an interactive session with --no-lock --no-clean
to examine results of jobs running with these.

The combining the new flags with the old ones and between each other is not limited
except for few cases which clearly exclude each other (--clean-only means only cleaning
and no cleaning and cleaning only don't make sense either). Although --no-lock and --no-clean
are meant to be used together, saying what is the correct or allowed use is hard because
the purpose of these flags is to disable the standard session behavior.