Add tests for clean and lock
Using Bash for the parallel and lock use case. It is little simpler than using multiprocessing
and even with that in place, it would be a good alternative test in the future because
it is a likely use case. It covers well the no locking part.

The no cleaning part is covered by in the Python test using two different approximations
of a real case. The current cleaning functionality is tested as well, but not in a standalone
test case nor with --clean-only.

The most relevant command line combinations are tested in a separate test.