CI: Use a specific Flake8 version
This adds a version specification to the Flake8 install in CI,
namely 3.8.0, the latest release.

It fixes string formatting issue (F523) in one file.
This fixes #619.

This also does some polish in the CI config file to tell that this is Python
and it removes the now unused continue-on-error lines.