libpython: Avoid race condition when reading region in use_temp_region()
When g.region is called without -u flag, it always adjusts and writes back the
computational region. When a parallel process is reading it too, it hits an empty
(or rarely corrupted) file.

In the provided test, this happens when reading the WIND file. At this point, the results
from the test must be expected manually.