Create maps and check their existence
Return codes of bg procs are not propagated, so something is needed to track processes.
Using maps created at the end of each Python script to do that.
Parametrize the Bash script to make it reusable also for a potential nesting region correctness test.
Also be more robust when the Python script is not available (zero returned on some platforms).