libpython: Avoid race condition when reading region in use_temp_region() (#638)
When g.region is called without -u flag, it always adjusts and writes back the
computational region. When a parallel process is reading it too, it hits an empty
(or rarely corrupted) file. region() and other region functions in core.py are already using it.

In the provided test, this happens when reading the WIND file. In the test, return codes
of background processes are not propagated (in Bash), so raster maps are created to track processes. These are created at the end of each Python script run.
The Bash script is parametrized to make it reusable also for a potential nesting region correctness test.

See also:
https://trac.osgeo.org/grass/ticket/2230 'g.region -p' writes new WIND file...
#250 r.import: ...WIND file disappears...
#630 unpredictable errors in the results with temp region (nesting suspected)