wxGUI/startup: Create grassdata automatically during the first GUI startup (#705)
New function for a grassdata directory in the default path for user's platform was created.

MS Windows is the special case and follows what the current installer does on Windows,
i.e., use Documents which should exist on all modern versions.
Linux, macOS etc. use home dir directly.

A temporary directory is used as a fallback for extreme, but real, cases when the home directory is not writable because this may happen in try-GRASS-GIS-online cases which is one of the cases when we want grassdata to be created automatically.  

Closes #682.