init: Start in last used mapset with --gui like --text
When the grass executable is started with --text, it uses the last used mapset.
Now also grass --gui behaves the same way and starts the main GUI without using the startup
window first to ask about the mapset.

The checks for starting in mapset are different because --text assumes it is interactive
and can just fail and give optins to the user while --gui needs to provide a GUI way.
As a result, --gui shows the statup window when the mapset is not usable for starting
(both use the ask-for-permission paradigm).