init: Start shell only with TTY
Start the (sub-)shell only when running in an interactive terminal (TTY).
Without checking this, the start fails in contexts when terminal is not available,
for example, running command from Alt+F2 dialog on many Linux distributions.
The same situation can be created using the nohup command.

A failure to retrieve shell PID from gisenv is no longer considered an error
for cases when shell was not started.

The interactive shell can be forced regadless of the TTY by using --text in the command line.