wxGUI/datacatalog: Add rename and delete of Location and Mapset in Datacatalog (#771)
Deleting and renaming is disabled for current mapset or location. Deleting is not possible. Renaming creates invalid (intermediate) state which g.mapset can't deal with if the renamed mapset is the current one.

General functions for create, rename, and delete mapset or location in guiutils that uses function from utils. Created LocationDialog and edited MapsetDialog.

Closes #710.

Co-authored-by: Anna Petrasova <kratochanna@gmail.com>
Co-authored-by: Vaclav Petras <wenzeslaus@gmail.com>