Improve messages and reload states
Preserve the original delete_mapset_interactively function with the db/loc/mapset parameters
as a wrapper of the new function.

Return value indicates change/modification. Error on the first mapset is no change. Error in the middle
is a change, but operation is interrupted, i.e. some possibly deleted and some not.