wxGUI/datacatalog: Allow delete on multiple mapsets in Data catalog (#795)
Add the function delete_mapsets_interactively() which is called after selecting more mapsets. Preserves the original
delete_mapset_interactively() function with the db/loc/mapset parameters as a wrapper of the new function. 

Checks ahead of time not to delete current or PERMANENT mapsets. Nothing is deleted if problem detected.
Lists all mapset with a problem (but not all problems for each mapset if more than one applies).

Return value indicates change/modification. Error on the first mapset is no change. Error in the middle
is a change, but operation is interrupted, i.e. some possibly deleted and some not.
Return value is geared towards updating (or not updating) displayed tree in data catalog and gives exact result,
so data catalog is reloaded only when one or more mapsets were (actually) deleted.
Message on error is more vague providing same info for all cases regardless of deleted mapsets
(assuming the reload of the tree and (in future better) checking ahead of time).

Exceptions/Errors are responsibility of the delete_mapsets_interactively() function.
General Exception try-except removed from the callers.

Co-authored-by: Vaclav Petras <wenzeslaus@gmail.com>