Remove the alternative mapset name part
Remove the alternative mapset name part as it does not make sense with how the part in gis_set.py is implemented. The function runs only when there is not db at all, so the location is always fresh, hence there is never an existing mapset (locked or unlocked). The implementation is not enough to handle the actual problem of locking anyway or existing startup location.