init: Start in a world (demo) location for first-time user with GUI (#868)
The first-time user, when GUI is used, will be brought into a newly created location with a empty dedicated mapset.
PERMANENT contains map of the world (countries). This should demonstrate correctly the mapset concept.
Consequently, the startup window is skipped for the first-time user and the main GUI with data tab is the first thing user sees.

When something goes wrong (e.g., installation does not include demolocation), the fallback is the startup window.
Limitations include missing only config dir or only grassdata (neither must be present) and empty grassdata.

The primary naming in the code is startup location because it is used primarily for startup
and at the same time it is more serious than demo. User sees a name based on the intended area of use and CRS.

Closes #862 and #897.