init: Change the history file according to the current mapset for Bash and GUI (#930)
Each new prompt checks the current mapset and when it is different, it write history (HISTFILE) to the old one, clears history list, and read history from the file in the new mapset. This is only for Bash.

There no need to warn about history for Bash in g.mapset anymore.

This also reloads history on change of current mapset in Console which at this point uses .bash_history too. Given the implementation of GConsoleWindow, even when the prompt/interactive console is hidden, giface needs to be passed, so the change needs to be present everywhere when GConsoleWindow is used. Direct usages of GPromptSTC could avoid the updates if history is not needed, but that's not implemented.

The old behavior was always there (with warning in g.mapset), but it became more important with always starting in a mapset as opposed to startup screen.