build: Case-insensitive check for GRASS in prefix (#949)
A check for the word grass in the installation directory (prefix) was lowercase only. This was too exclusive and caused a warning and unnecessary interaction for cases where GRASS is written in uppercase which is the case, e.g., on macOS with /Applications/GRASS-7.9.app/...

This checks against a lower-cased install directory path, hence making it case-insensitive.