r.geomorphon: Put the FORMS enum right
Dàibhidh Grannd (English Language and Linguistics, School of Critical
Studies, University of Glasgow) points out that in the plugin HTML
documentation and in the original paper landform category 5 stands for
"SP" (spur), not "CV" (convex slope), and landform category 7 stands
for "HL" (hollow), not "CN" (concave slope). Amend that in the code and
update the comments.

Use the type where it belongs and replace a number of hard-coded
constants with enumeration constants. Reduce a variable scope.