r.geomorphon: Fix shape calculations
The plugin sometimes produced apparently invalid (several orders of
magnitude off or "not a number") values for elongation and width. Add
missing initializations to shape() in geom.c to fix that and the azimuth
calculation as well: when finding a sum of values, initialize the
holding variable with 0; when finding a minimum or a maximum for a set
of values, initialize the holding variable with the first value.