r.geomorphon: Untangle other min/max code
In exposition() and range() the hard-coded constants delivered the right
result for the data, but it was not immediately obvious, so just
initialize the holding variable with the first value for clarity. In
exposition() rewrite the assignment so it does not look like a standard
maximum expression, because it is not one.