r.geomorphon: Use o_forms where it belongs
The for loops that iterate over opt_output[] started from hard-coded 1
(o_forms), which is the first output, but not the first element in the
array. Use the enumeration constant to iterate properly regardless of
the amount of any preceding items in the array. Remove a comment that no
longer applies.