wxGUI/mapdisp: Overlay properties in context menu (#1077)
Properties of legends and other overlays are now also available from the context
menu on right click (not just double click). So, you don't need to know about the
double click to edit them.

Tooltip is simplified, more general, and does not advertise the double click anymore
resulting in shorter, easier to read text.

Raster legend resize is renamed to resize and move because that's what it is.