g.copy: Report failed copy using an error message
Changes in behavior:

1. Non-zero return code (EXIT_FAILURE) has now also a correspoding
   error message (ERROR:) not just warning messages (from
   do_copy.c).
2. Skipping an input as ineligible for copying may now result in
   an error (fatal error with EXIT_FAILURE). The error newly occures when
   all inputs are skipped or, most importantly, when only one input
   was provided and it was skipped.

The skipping (in main.c) is now consistently reported as skipping.
The warning messages also contain more details and the new error
messages point to them.

The M_do_copy function (do_copy.c) now starts with 'Copying' because
it is at the begining of the process instead of 'Copy' which is misleading
(is the user supposed to copy).

The user manual has now a new intro and newly describes the behavior on
error discussing the much simpler single input case separatelly.