rasterlib: Explain missing header file in get_cellhd (#933)
When header file is missing, the user is presented with 'Unable to open header file for raster map'.
Clearly not the same as unable to open map, but it is not clear what is wrong as it is not clear
what actually is the relation ship of the header file to the raster map or why this error occurred
and not the map not found one.

This adds some hints on what may have happened, so the user can start investigation (or recomputation) being
more informed on what might be the problem. At the same time, this keeps the original (core of the) message there
so that there is no doubt about what was the immediate cause of the error.

This PR also fixes the construction of an English sentence in the translatable string and replaces that by
adding another sentence.

The normal case and the two reclass cases each give a different info as the context is likely different.
In all cases 'It seems' and 'Perhaps' are used to indicate uncertainty of the message.

The documentation of the function was outdated mentioning return codes and warnings, now updated with G_fatal_error().

Since most of the lines in this file change, the indent script was applied.