Small code fixes for d.legend and Cairo_Text (#1076)
MaxLabel needs to hold buff plus one (e.g. for a space) to avoid several
warnings: sprintf may write a terminating nul past the end of the destination
[-Wformat-overflow=] with GCC.

D_save_command needs to be called only one per standard module call
(although should be removed in the future). draw() is called twice with bg drawing.

D_text call seems to be badly indented, but in a correct branch, so fixing
warning: this if clause does not guard... [-Wmisleading-indentation].

Various iconv documetations specify the retun value on error as (iconv_t) -1
for iconv_open, so test specifically for that and avoid
warning: ordered comparison of pointer with integer zero [-Wextra].

Fix also a typo and empty line at the end of file.