d.legend: Fix position of units and background (#1075)
Do the same for units from metadata (r.info, r.support units=, aka units_bottom)
as for the title, i.e., use the same computation. Use of the right variables
(not title or (the other) units). Use the same check already used for title
for the units too when drawing the background.
(This is applicable only to the do_smooth if statement branch,
not the non FP, no smoothing branch in the code.)